pyproject.toml
[tool.poetry]
name = "crud-basico"
version = "0.1.0"
description = ""
authors = ["PauloHBSF <paulohbsfilho@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.34"
pydantic = "^2.9.1"
psycopg2 = "^2.9.9"
dotenv-python = "^0.0.1"
python-dotenv = "^1.0.1"
fastapi = {extras = ["standard"], version = "^0.114.0"}
psycopg = {extras = ["binary"], version = "^3.2.1"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
